Split packet parser module
